Use pages for notebooks
