+++ demo/package.json
@@ -18,7 +18,8 @@
   "dependencies": {
     "jupyter-iframe-commands-host": "^0.1.0",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "react-router": "^7.3.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.19.0",
